rename simple measurement to avoid confussion
