+++ package/PartSeg/plugins/napari_widgets/__init__.py
@@ -6,7 +6,7 @@ from PartSeg.plugins.napari_widgets.roi_extraction_algorithms import ROIAnalysis
 
 @napari_hook_implementation
 def napari_experimental_provide_dock_widget():
-    from PartSeg.plugins.napari_widgets.measurement_widget import SimpleMeasurement
+    from PartSeg.plugins.napari_widgets.simple_measurement_widget import SimpleMeasurement
 
     return SimpleMeasurement
