+++ package/PartSeg/plugins/napari_widgets/simple_measurement_widget.py
