+++ package/tests/test_PartSeg/test_napari_widgets.py
@@ -132,7 +132,7 @@ def test_profile_preview_dialog(part_settings, register, qtbot, monkeypatch, tmp
 
 @napari_skip
 def test_measurement_create(make_napari_viewer, qtbot):
-    from PartSeg.plugins.napari_widgets.measurement_widget import SimpleMeasurement
+    from PartSeg.plugins.napari_widgets.simple_measurement_widget import SimpleMeasurement
 
     data = np.zeros((10, 10), dtype=np.uint8)
